Added: Cookie login listener
Since LDAP and the site aren't using the same database, we need to keep
the database in sync with LDAP. I have thought a lot about this, and to
do this I decided to hook it up to the Authentication event of laravel.
This is getting triggered every time you log in, AND every time laravel
uses a cookie to log in. This means we can also log the user out
automatically if it couldn't be found in the LDAP database. Using this
method means we won't make an LDAP connection every time a user loads a
page.