Added: Vegetarian checkbox when signin in
This is because this is an option that is chosen a lot, so why not make
a checkbox for it.

Also removed the whole Yes No columns for cooks and diswhasher from the
overview and replaced them with labels. In my opinion it is useless to
create an entire column when it's only applicable for just 2 or 3 users.