Added: API for getting all FAQ items
Also took the liberty to add a convention that all models that will be sent to the user to be of the form ...ResponseModel. And if we ever retrieve objects as models from the user it will be of name ...RequestModel. This to make a clear distinction between our database models and our response models.